pyproject.toml
[tool.poetry]
name = "ct-finals"
version = "0.1.0"
description = ""
authors = ["Chenghsuan Ho <chengxuanjhe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.5"
PySpice = "^1.5"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
